Backend/requirements.txt
django==3.2
djangorestframework==3.14.0
psycopg2==2.9.5
drf-flex-fields==1.0.0
djangorestframework-simplejwt==5.2.2
django-filter==22.1